% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BARModel.R
\name{BayesRankCovSimp}
\alias{BayesRankCovSimp}
\title{Bayesian Analysis of Rank data with entities' Covariates}
\usage{
BayesRankCovSimp(
  pair.comp.ten,
  X.mat = matrix(NA, nrow = dim(pair.comp.ten)[1], ncol = 0),
  tau2.alpha = 5^2,
  nu.alpha = 3,
  tau2.beta = 5^2,
  nu.beta = 3,
  n.item = dim(pair.comp.ten)[1],
  n.ranker = dim(pair.comp.ten)[3],
  p.cov = ncol(X.mat),
  iter.max = 5000,
  para.expan = TRUE,
  print.opt = 100,
  initial.list = NULL
)
}
\arguments{
\item{X.mat}{An \eqn{N} by \eqn{L} covariate matrix for the \eqn{N} entities with \eqn{L} covariates.}

\item{tau2.alpha}{The scale parameter for the scaled inverse chi-squared prior on \eqn{\sigma^2_alpha}.}

\item{nu.alpha}{The d.f. for the scaled inverse chi-squared prior on \eqn{\sigma^2_alpha}.}

\item{tau2.beta}{The scale parameter for the scaled inverse chi-squared prior on \eqn{\sigma^2_beta}.}

\item{nu.beta}{The d.f. for the scaled inverse chi-squared prior on \eqn{\sigma^2_beta}.}

\item{iter.max}{Number of iterations for Gibbs sampler.}

\item{para.expan}{Logical variable for whether using parameter expansion in the Gibbs sampler.}

\item{pair.com.ten}{An \eqn{N} by \eqn{N} by \eqn{M} pairwise comparison tensor for all \eqn{N} entities and \eqn{M} rankers, where the (\eqn{i},\eqn{j},\eqn{m}) element equals 1 if \eqn{i} is ranked higher than \eqn{j} by ranker \eqn{m}, 0 if \eqn{i} is ranker lower than \eqn{j}, and NA if the relation between \eqn{i} and \eqn{j} is missing. Note that the diagonal elements (\eqn{i},\eqn{i},\eqn{m})'s for all rankers should be set to NA as well.}
}
\value{
A list containing posterior samples of all the missing evaluation scores for all rankers and all the model parameters.
}
\description{
Implement the Bayesian model for rand data with ranked entities' covariates information.
}
