% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankPairComp.R
\name{FullRankToPairComp}
\alias{FullRankToPairComp}
\title{Compute Pairwise Comparison Matrix for Full Ranking List of the Entities}
\usage{
FullRankToPairComp(rank.vec, n = length(rank.vec))
}
\arguments{
\item{rank.vec}{A full ranking list containing the ranks of all the \eqn{N} entities. Note that here we follow the usual definition of rank in R, that is, the larger the evaluation score of an entity, the larger this entity's rank is. Specifically, for a full ranking list of \eqn{N} entities, the rank of an entity equals \eqn{N+1} minus its ranked position.}
}
\value{
An \eqn{N} by \eqn{N} pairwise comparison for all \eqn{N} entities, where the (\eqn{i},\eqn{j}) element equals 1 if \eqn{i} is ranked higher than \eqn{j}, and 0 if \eqn{i} is ranker lower than \eqn{j}. Note that the diagonal elements (\eqn{i},\eqn{i})'s are set to NA.
}
\description{
Compute the pairwise comparison matrix from the ranking lists of the ranked entities.
}
