% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankPairComp.R
\name{PartRankToPairComp}
\alias{PartRankToPairComp}
\title{Compute Pairwise Comparison Matrix for Partial Ranking Lists of the Entities}
\usage{
PartRankToPairComp(rank.matrix, n = nrow(rank.matrix), overlap = TRUE)
}
\arguments{
\item{rank.matrix}{An \eqn{N} by \eqn{K} matrix containing \eqn{K} partial ranking lists for the \eqn{N} ranked entities. Here each column of \code{rank.matrix} contains a partial ranking list for the \eqn{N} entities, where the value of an entity is NA if its pairwise comparison information is missing in this partial list. Note that here we follow the usual definition of rank in R, that is, the larger the evaluation score of an entity, the larger this entity's rank is. Specifically, for a full ranking list of \eqn{N} entities, the rank of an entity equals \eqn{N+1} minus its ranked position.}
}
\value{
An \eqn{N} by \eqn{N} pairwise comparison for all \eqn{N} entities, where the (\eqn{i},\eqn{j}) element equals 1 if \eqn{i} is ranked higher than \eqn{j}, 0 if \eqn{i} is ranker lower than \eqn{j}, and NA if the relation between \eqn{i} and \eqn{j} is missing. Note that the diagonal elements (\eqn{i},\eqn{i})'s are set to NA.
}
\description{
Compute the pairwise comparison matrix from partial ranking lists of the ranked entities.
}
