% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankPairComp.R
\name{RankDist}
\alias{RankDist}
\title{Compute the Normalized Kendall Tau Distance from two Full Ranking Lists}
\usage{
RankDist(Rank1, Rank2, method = "kendall")
}
\arguments{
\item{Rank1}{A full ranking list containing the ranks of all the \eqn{N} entities. Note that here we follow the usual definition of rank in R, that is, the larger the evaluation score of an entity, the larger this entity's rank is. Specifically, for a full ranking list of \eqn{N} entities, the rank of an entity equals \eqn{N+1} minus its ranked position.}

\item{Rank2}{Another full ranking list containing the ranks of all the \eqn{N} entities.}
}
\value{
The normalized Kendall tau distance from the two full ranking lists.
}
\description{
Compute the Normalized Kendall Tau Distance from two Full Ranking Lists
}
