% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankPairComp.R
\name{RankToplist}
\alias{RankToplist}
\title{Compute Ranks, Top List and Evaluation Scores from each other}
\usage{
RankToplist(Rank = NULL, Toplist = NULL, mu = NULL)
}
\arguments{
\item{Rank}{A full ranking list containing the ranks of all the \eqn{N} ranked entities. Note that here we follow the usual definition of rank in R, that is, the larger the evaluation score of an entity, the larger this entity's rank is. Specifically, for a full ranking list of \eqn{N} entities, the rank of an entity equals \eqn{N+1} minus its ranked position.}

\item{Toplist}{A ordered vector listing the indices of the \eqn{N} entities with decreasing ranks or evaluation scores.}

\item{mu}{A vector of the evaluation scores for all the \eqn{N} ranked entities.}
}
\value{
A list containing the ranks, top list and evaluation scores for all the \eqn{N} entities.
}
\description{
Compute Ranks, Top List and Evaluation Scores from each other
}
